# Chunk generation optimizations