--- a/pumpkin-world/src/generation/proto_chunk.rs
+++ b/pumpkin-world/src/generation/proto_chunk.rs
@@ -579,6 +579,7 @@ impl<'a> ProtoChunk<'a> {
     }
 
     pub fn get_biome_for_terrain_gen(&self, global_block_pos: &Vector3<i32>) -> &'static Biome {
+        // TODO: See if we can cache this value
         let seed_biome_pos = biome::get_biome_blend(
             self.bottom_y(),
             self.height(),
@@ -628,8 +629,8 @@ impl<'a> ProtoChunk<'a> {
                     self.terrain_cache
                         .terrain_builder
                         .place_badlands_pillar(self, x, z, top_block);
-                    // Get the top block again if we placed a pillar!
 
+                    // Get the top block again if we placed a pillar!
                     top_block = self.top_block_height_exclusive(&Vector2::new(local_x, local_z));
                 }