<!-- Empty or bad Descriptions are not welcome, Don't waste my time -->

## Description
Shaves a couple of milliseconds off of each chunk during the noise generation phase

## Testing

Please follow our [Coding Guidelines](https://github.com/Pumpkin-MC/Pumpkin/blob/master/CONTRIBUTING.md#coding-guidelines)