@kralverde 2025-09-14
Old (with new benchmark):
`time:   [59.816 ms 59.850 ms 59.885 ms]`

New:
`time:   [54.917 ms 54.976 ms 55.033 ms]`